requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
pandas
pytest
httpx
python-dotenv
